+++ mm2src/coins/lp_coins.rs
@@ -240,7 +240,6 @@ use coin_errors::{MyAddressError, ValidatePaymentError, ValidatePaymentFut, Vali
 pub mod coins_tests;
 
 pub mod eth;
-use eth::eth_swap_v2::{PaymentStatusErr, ValidatePaymentV2Err};
 use eth::GetValidEthWithdrawAddError;
 use eth::{eth_coin_from_conf_and_request, get_eth_address, EthCoin, EthGasDetailsErr, EthTxFeeDetails,
           GetEthAddressError, SignedEthTx};
@@ -320,7 +319,8 @@ pub mod z_coin;
 use crate::coin_balance::{BalanceObjectOps, HDWalletBalanceObject};
 use z_coin::{ZCoin, ZcoinProtocolInfo};
 #[cfg(feature = "enable-sia")] pub mod sia;
-use crate::eth::eth_swap_v2::PrepareTxDataError;
+use eth::eth_swap_v2::PrepareTxDataError;
+use eth::eth_swap_v2::{PaymentStatusErr, ValidatePaymentV2Err};
 #[cfg(feature = "enable-sia")] use sia::SiaCoin;
 
 pub type TransactionFut = Box<dyn Future<Item = TransactionEnum, Error = TransactionErr> + Send>;
